refactor(runtime): separate commit and abort protocols
